#84 move 100 lines of "util" code into dedicated safetyVerifier module, cleanup, comments
